refactor(v0.6): split core/lifecycle/replay_graph.py 23.0KB → 4-file package
CLAUDE.md rule #5: no file in core/ exceeds 20 KB. The legacy single
file (23.0 KB) sits above the cap. This split into a package preserves
the public + private import surface byte-identically — every caller
(routes/admin.py + the T5 test suite) keeps working unchanged.

Sub-files (all < 20 KB):
  __init__.py     3.7 KB   re-exports (façade) + sqlite3 re-export
                            for the I1 audit-only invariant monkeypatch
  snapshot.py     2.7 KB   GraphSnapshot dataclass + _empty_snapshot
  handlers.py     7.8 KB   9 _h_* per-event handlers + _HANDLERS dict
                            + LIFECYCLE_EVENT_TYPES sanity assert
  db_read.py      3.2 KB   _default_db_path + _read_lifecycle_events
                            (the only audit-log side-channel)
  primitives.py   9.2 KB   reconstruct_graph_at + view_from_snapshot
                            + _validity_contains

Quality delta: exempt (label: chore) — pure refactor, no behaviour
change. All 55 T5 + RAB + replay tests stay green; the new lock-test
tests/test_v06_replay_graph_module_size.py (8 assertions) pins the
20 KB cap + the public/private import surface + the LIFECYCLE_EVENT
taxonomy invariant.

The one repo-cwd-sensitive change is _default_db_path: the legacy file
walked 3 dirname() levels up from core/lifecycle/replay_graph.py to
reach the repo root. Now at core/lifecycle/replay_graph/db_read.py
the package sits one level deeper, so the new path walks 4 levels —
verified by the T5 fresh-DB tests which exercise the default path
resolution.

Closes #N/A (CLAUDE.md rule #5 hygiene — refactor 3/5).
